type conversion

+++ src/cpp/src/continuous_batching/model_runner.hpp
@@ -732,7 +732,7 @@ public:
                                             "tree_position_ids[", token_id, "] must be non-negative, got ", tree_pos_id,
                                             "; position_ids_idx=", position_ids_idx,
                                             ", seq_id=", sequence->get_id());
-                            position_ids_data[position_ids_idx] = group_position_id + tree_pos_id;
+                            position_ids_data[position_ids_idx] = static_cast<int64_t>(group_position_id) + tree_pos_id;
                         } else {
                             position_ids_data[position_ids_idx] = position_id;
                         }
